refactor: move task service type to new types folder
